py/src/requirements.txt
pyzmq==15.2.0
